% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_bootstrap.R
\name{boot_indices_internal}
\alias{boot_indices_internal}
\alias{boot_indices_internal.sample}
\alias{boot_indices_internal.both}
\title{Bootstrap indices}
\usage{
boot_indices_internal(...)

\method{boot_indices_internal}{sample}(obj_boot, n_boot = 50, seed = 1, ...)

\method{boot_indices_internal}{both}(obj_boot, n_boot = 50, seed = 1, ...)
}
\arguments{
\item{...}{Additional parameters}

\item{obj_boot}{Object to estimate hte}

\item{n_boot}{Number of bootstrap samples}

\item{seed}{Seed to run simulations.}
}
\value{
List with following parameters:
\item{ind_population}{Vector with population indices}
\item{ind_sample}{Vector with sample indices}
}
\description{
Internal generic function to obtain bootstrap indices
}
\section{Methods (by class)}{
\itemize{
\item \code{sample}: Obtain bootstrap indices bootstrapping withing the sample

\item \code{both}: Obtain bootstrap indices bootstrapping withing the sample
}}

