% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_bootstrap.R
\name{bootstrap_indices}
\alias{bootstrap_indices}
\title{Bootstrap indices from the sample}
\usage{
bootstrap_indices(sample_sizes, n_boot = 50, seed = 1)
}
\arguments{
\item{sample_sizes}{Data from the sample}

\item{n_boot}{Number of bootstrap samples}

\item{seed}{Seed to run simulations.}
}
\value{
List with following parameters:
\item{ind_population}{Vector with population indices}
\item{ind_sample}{Vector with sample indices}
}
\description{
Bootstrap indices from the sample
}
