% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{estimate_hte}
\alias{estimate_hte}
\alias{estimate_hte.OR}
\alias{estimate_hte.IPW}
\alias{estimate_hte.NIPW}
\alias{estimate_hte.AIPW}
\title{Estimate hte}
\usage{
estimate_hte(...)

\method{estimate_hte}{OR}(obj_hte, params_OR, params_bootstrap, ...)

\method{estimate_hte}{IPW}(
  obj_hte,
  params_p_score,
  params_impute_y,
  estimated_p_score,
  params_bootstrap,
  ...
)

\method{estimate_hte}{NIPW}(
  obj_hte,
  params_p_score,
  params_impute_y,
  estimated_p_score,
  params_bootstrap,
  ...
)

\method{estimate_hte}{AIPW}(
  obj_hte,
  params_p_score,
  params_impute_y,
  params_OR,
  estimated_p_score,
  params_bootstrap,
  ...
)
}
\arguments{
\item{...}{Additional parameters}

\item{obj_hte}{Object to estimate hte}

\item{params_OR}{List with parameters to fit outcome regression model:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item clust_RF - add clustering to random forest,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
\item type_model - type of outcome.
}}

\item{params_bootstrap}{List with parameters to obtain bootstrap variance:
\itemize{
\item boot_var = FALSE,
\item n_boot = 500,
\item boot_seed = 10,
}}

\item{params_p_score}{List with parameters to fit propensity score:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item clust_RF - add clustering to random forest,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
}}

\item{params_impute_y}{List with parameters to fit imputation model:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item clust_RF - add clustering to random forest,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
\item type_model - type of outcome.
}}

\item{estimated_p_score}{Vector of estimated propensity scores, default: \code{estimated_p_score = NULL}.}
}
\value{
List with following parameters:
\item{y_hat_out_of_sample}{imputed out of sample outcomes}
\item{y_hat_sample}{predicted sample data using fitted model}
\item{outcome_fit}{fitted model to impute the data}
}
\description{
Internal generic function to estimate hte
}
\section{Methods (by class)}{
\itemize{
\item \code{estimate_hte(OR)}: Estimate heterogeneous treatment effects using OR

\item \code{estimate_hte(IPW)}: Estimate heterogeneous treatment effects using IPW

\item \code{estimate_hte(NIPW)}: Estimate heterogeneous treatment effects using NIPW

\item \code{estimate_hte(AIPW)}: Estimate heterogeneous treatment effects using AIPW

}}
