% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{fct_form}
\alias{fct_form}
\alias{fct_form.sin}
\alias{fct_form.exp}
\alias{fct_form.sin_exp}
\alias{fct_form.lin}
\title{Generate one population}
\usage{
fct_form(...)

\method{fct_form}{sin}(fct_obj, ...)

\method{fct_form}{exp}(fct_obj, ...)

\method{fct_form}{sin_exp}(fct_obj, ...)

\method{fct_form}{lin}(fct_obj, ...)
}
\arguments{
\item{...}{Additional parameters}

\item{fct_obj}{Function form object}
}
\value{
A data vector
}
\description{
This is a generic function to generate outcome variables
}
\section{Methods (by class)}{
\itemize{
\item \code{fct_form(sin)}: Sine functional form

\item \code{fct_form(exp)}: Sine functional form

\item \code{fct_form(sin_exp)}: Sine(exp) functional form

\item \code{fct_form(lin)}: Linear functional form

}}
