% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tau.R
\name{fct_treat_untreat}
\alias{fct_treat_untreat}
\title{Internal function to compute group means for the treated
and untreated}
\usage{
fct_treat_untreat(subpopulation, type_tau)
}
\arguments{
\item{subpopulation}{Subpopulation to compute causal effect}

\item{type_tau}{Type of computation, \code{regular}: constant weights,
\code{adaptive}: adaptive weights}
}
\description{
Internal function to compute group means for the treated
and untreated
}
