% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{fit_p_score}
\alias{fit_p_score}
\title{Fit propensity score}
\usage{
fit_p_score(obj_hte, params_p_score)
}
\arguments{
\item{params_p_score}{List with parameters to fit propensity score:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
}}
}
\value{
Estimates of propensity score.
}
\description{
Internal function to obtain subpopulation heterogenous treatment effects
}
