% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_y.R
\name{fit_y}
\alias{fit_y}
\alias{fit_y.EBLUP}
\alias{fit_y.MQ}
\alias{fit_y.RF}
\alias{fit_y.XGB}
\title{Geeneric function to fit values of y}
\usage{
fit_y(...)

\method{fit_y}{EBLUP}(obj_fit_y, type_model, params_bootstrap, ...)

\method{fit_y}{MQ}(obj_fit_y, type_model, ...)

\method{fit_y}{RF}(obj_fit_y, type_model, tune_RF, ...)

\method{fit_y}{XGB}(obj_fit_y, type_model, xgboost_params, ...)
}
\arguments{
\item{...}{Additional parameters.}

\item{obj_fit_y}{Object to impute data}

\item{type_model}{Type of outcome.}

\item{params_bootstrap}{List with parameters to obtain bootstrap variance:
\itemize{
\item boot_var = FALSE,
\item n_boot = 500,
\item boot_seed = 10,
}}

\item{tune_RF}{Tune parameters in random forest? Default = FALSE.}

\item{xgboost_params}{List with parameters to run xgboost:
\itemize{
\item CV_XGB - logical variable, use cross-validation for gradient boosting? Default: TRUE.
\item nfolds - number of folds in cross-validation, default: nfolds = 5.
\item nrounds - the max number of iterations, default: nrounds = 50.
}}
}
\value{
List with following parameters:
\item{y_hat_out_of_sample}{imputed out of sample outcomes}
\item{y_hat_sample}{predicted sample data using fitted model}
\item{outcome_fit}{fitted model to impute the data}
}
\description{
Geeneric function to fit values of y
}
\section{Methods (by class)}{
\itemize{
\item \code{EBLUP}: Function to estimate fit_y using EBLUP

\item \code{MQ}: Function to estimate fit_y using MQ

\item \code{RF}: Function to estimate fit_y using RF

\item \code{XGB}: Function to estimate fit_y using RF
}}

