% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{gen_outcome}
\alias{gen_outcome}
\alias{gen_outcome.continuous}
\alias{gen_outcome.binary}
\alias{gen_outcome.poisson}
\title{Generate one population}
\usage{
gen_outcome(...)

\method{gen_outcome}{continuous}(
  regression_obj,
  Xreg_outcome,
  errors_outcome,
  coeffs,
  A,
  Ni,
  m,
  rand_eff_outcome,
  start_seed = 1,
  ...
)

\method{gen_outcome}{binary}(
  regression_obj,
  Xreg_outcome,
  errors_outcome,
  coeffs,
  A,
  Ni,
  m,
  rand_eff_outcome,
  start_seed = 1,
  ...
)

\method{gen_outcome}{poisson}(
  regression_obj,
  Xreg_outcome,
  errors_outcome,
  coeffs,
  A,
  Ni,
  m,
  rand_eff_outcome,
  start_seed = 1,
  ...
)
}
\arguments{
\item{...}{Additional parameters}

\item{regression_obj}{Object defining regression type}

\item{Xreg_outcome}{Outcome covariates times coefficients \code{X beta} (without intercept)}

\item{errors_outcome}{List of parameters to generate errors, that is:
\itemize{
\item var_e
\item mean_e
\item frac_out
\item var_e_out
\item mean_e_out
\item disturbance_outcome
}}

\item{coeffs}{List of coefficients for generating outcomes and propensity scores, tat is:
\itemize{
\item intercept_outcome
\item intercept_p_score
\item coef_outcome
\item coef_p_score
\item mean_A
\item var_A
}}

\item{A}{Treatment status}

\item{Ni}{Number of units in areas}

\item{m}{Number of subpopulations}

\item{rand_eff_outcome}{List of parameters to generate random effects for outcome model, that is:
\itemize{
\item var_re
\item mean_re
\item frac_out
\item var_re_out
\item mean_re_out
}}

\item{start_seed}{Seed to replicate simulations}
}
\value{
A data frame with generated population
}
\description{
This is a generic function to generate outcome variables
}
\section{Methods (by class)}{
\itemize{
\item \code{continuous}: Generate population with normally distributed outcomes

\item \code{binary}: Generate population with Poisson distributed outcomes

\item \code{poisson}: Generate population with binomial distributed outcomes
}}

