% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample.R
\name{generate_one_sample}
\alias{generate_one_sample}
\title{Generate one sample from population}
\usage{
generate_one_sample(
  population,
  size_samples,
  sample_part = "sampled",
  get_index = F
)
}
\arguments{
\item{population}{Data frame with a generated population}

\item{size_samples}{Sample sizes in each subpopulation/area}

\item{sample_part}{Select sampled, non_sampled or both parts
or population, default: sample_part = sampled}

\item{get_index}{Get the index of sampled observations}
}
\description{
Generate one sample from population
}
