% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rv_out.R
\name{generate_poisson}
\alias{generate_poisson}
\title{Generate Poisson distributed rv with outliers}
\usage{
generate_poisson(n = 10, mu = 1, disturbance = 5, frac_out = 0, seed = 1)
}
\arguments{
\item{n}{Number of errors (residuals)}

\item{mu}{Mean of Poisson rv}

\item{disturbance}{Shift parameter to generate outliers}

\item{frac_out}{Fraction of outlying random effects}

\item{seed}{Seed to replicate simulations}
}
\value{
Vector of random effects with outliers.
}
\description{
This is an internal function to generate errors from
a Poisson distribution with outliers. It is assumed that the parameters
passed into the function are already checked, so we do not recommend to
use this function separately.
}
