% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rv_out.R
\name{generate_re}
\alias{generate_re}
\title{Generate random elements}
\usage{
generate_re(
  n,
  type_re = c("random_effects", "errors"),
  var_re,
  mean_re,
  frac_out,
  var_re_out,
  mean_re_out,
  seed = 1
)
}
\arguments{
\item{n}{Number of random effects.}

\item{type_re}{Type of the random elements: random_effects, errors.}

\item{var_re}{Variance of random effects.}

\item{mean_re}{Mean of random effects.}

\item{frac_out}{Fraction of outlying random effects.}

\item{var_re_out}{Variance of outlying random effects.}

\item{mean_re_out}{Mean of outlying random effects.}

\item{seed}{Seed to replicate simulations.}
}
\value{
Vector of random effects or errors with outliers.
}
\description{
This is an internal function to generate random effects or errors from
a normal distribution with outliers. It is assumed that the parameters
passed into the function are already checked, so we do not recommend to
use this function separately.
}
\examples{

random_effects <- generate_re(n = 100,
type_re = "random_effects",
var_re = 2,
mean_re = 0,
frac_out = 0.05,
var_re_out = 20,
mean_re_out = 0)


random_effects <- generate_re(n = 100,
type_re = "errors",
var_re = 2,
mean_re = 0,
frac_out = 0.05,
var_re_out = 20,
mean_re_out = 0)


}
