% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample.R
\name{generate_sample}
\alias{generate_sample}
\title{Sample from the population}
\usage{
generate_sample(
  populations,
  ni_size = 5,
  sample_part = "sampled",
  get_index = F
)
}
\arguments{
\item{populations}{List with populations}

\item{ni_size}{Sample size}

\item{sample_part}{Select sampled, non_sampled or both parts}

\item{get_index}{Get the index of sampled observations
or population, default: sample_part = sampled}
}
\description{
Sample from the population
}
