% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X.R
\name{generate_X}
\alias{generate_X}
\title{Generate covariates}
\usage{
generate_X(
  n,
  p = 2,
  covariance_norm = NULL,
  cov_type = c("norm", "lognorm", "unif"),
  seed = 1
)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of coefficients.}

\item{covariance_norm}{Variance-covariance matrix for
correlated normally distributed covariates.}

\item{cov_type}{Type of coefficients.}

\item{seed}{Seed to run simulations.}
}
\value{
Matrix of generated covariates.
}
\description{
Wrapper function to generate covariates
}
\examples{

# Generate X from normal distribution
X_norm <- generate_X(
n = 10,
p = 2,
covariance_norm = NULL,
cov_type = c("norm"),
              seed = 1)


# Generate X from log-normal distribution
X_norm <- generate_X(
n = 10,
p = 2,
covariance_norm = NULL,
cov_type = c("lognorm"),
              seed = 1)


# Generate X from uniform distribution
X_norm <- generate_X(
n = 10,
p = 2,
covariance_norm = NULL,
cov_type = c("unif"),
              seed = 1)



}
