% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{get_default_errors_outcome}
\alias{get_default_errors_outcome}
\title{Get default parameters to generate outcomes with possible disturbances}
\usage{
get_default_errors_outcome()
}
\description{
Get default parameters to generate outcomes with possible disturbances
}
