% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{get_default_rand_eff_outcome}
\alias{get_default_rand_eff_outcome}
\title{Get default parameters to generate random effects with possible disturbances for outcome model}
\usage{
get_default_rand_eff_outcome()
}
\description{
Get default parameters to generate random effects with possible disturbances for outcome model
}
