% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{get_default_rand_eff_p_score}
\alias{get_default_rand_eff_p_score}
\title{Get default parameters to generate random effects with possible disturbances for propensity score model}
\usage{
get_default_rand_eff_p_score()
}
\description{
Get default parameters to generate random effects with possible disturbances for propensity score model
}
