% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X.R
\name{get_X}
\alias{get_X}
\alias{get_X.norm}
\alias{get_X.lognorm}
\alias{get_X.unif}
\title{Get covariate \code{X}}
\usage{
get_X(...)

\method{get_X}{norm}(cov_type, n, p, covariance_norm, seed, ...)

\method{get_X}{lognorm}(cov_type, n, p, covariance_norm, seed, ...)

\method{get_X}{unif}(cov_type, n, p, covariance_norm, seed, ...)
}
\arguments{
\item{...}{Additional parameters}

\item{cov_type}{Type of coefficients.}

\item{n}{Number of observations.}

\item{p}{Number of coefficients.}

\item{covariance_norm}{Variance-covariance matrix for
correlated normally distributed covariates.}

\item{seed}{Seed to run simulations.}
}
\value{
A data frame with generated population
}
\description{
Generic function to generate covariate \code{X} from normal, log-normal or uniform distribution
}
\section{Methods (by class)}{
\itemize{
\item \code{get_X(norm)}: Generate covariate from normal distribution.

\item \code{get_X(lognorm)}: Generate covariate from log-normal distribution.

\item \code{get_X(unif)}: Generate covariate from uniform distribution.

}}
