% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{hte}
\alias{hte}
\title{Heterogenous treatment effect}
\usage{
hte(
  formula_y,
  formula_p_score,
  type_tau,
  data_sample,
  data_out_of_sample,
  method_y = c("EBLUP", "MQ", "RF"),
  method_p_score = c("EBLUP", "MQ", "RF"),
  p_score_estimate = NULL,
  tune_RF = F,
  ...
)
}
\arguments{
\item{formula_y}{Formula for fitting outcome regression}

\item{formula_p_score}{Formula for fitting outcome regression}

\item{type_tau}{Type of computation, \code{regular}: constant weights,
\code{adaptive}: adaptive weights}

\item{data_sample}{Data from the sample}

\item{data_out_of_sample}{Data from the population}

\item{method_y}{Method to predict/fit the outcome}

\item{method_p_score}{Method to predict/fit propnsity score}

\item{p_score_estimate}{Vector of estimated propensity scores}

\item{tune_RF}{Tune random forest parameters}

\item{...}{Additional parameters}
}
\description{
Calculates heterogrenous treatment effects for subpopulations
}
