% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R, R/impute_y.R
\name{estimate_hte.OR}
\alias{estimate_hte.OR}
\alias{estimate_hte.IPW}
\alias{imputation_y}
\alias{imputation_y.EBLUP}
\alias{imputation_y.MQ}
\alias{imputation_y.RF}
\alias{imputation_y.XGB}
\title{Internal generic function to impute values of y}
\usage{
\method{estimate_hte}{OR}(obj_hte, params_OR, ...)

\method{estimate_hte}{IPW}(obj_hte, params_p_score, params_impute_y, ...)

imputation_y(...)

\method{imputation_y}{EBLUP}(obj_imputation_y, model_formula, type_model, ...)

\method{imputation_y}{MQ}(obj_imputation_y, model_formula, type_model, ...)

\method{imputation_y}{RF}(obj_imputation_y, model_formula, type_model, tune_RF, ...)

\method{imputation_y}{XGB}(obj_imputation_y, model_formula, type_model, xgboost_params, ...)
}
\arguments{
\item{...}{Additional parameters.}

\item{obj_imputation_y}{Object to impute data}

\item{model_formula}{Model formula.}

\item{type_model}{Type of outcome.}

\item{tune_RF}{Tune parameters in random forest? Default = FALSE.}

\item{xgboost_params}{List with parameters to run xgboost:
\itemize{
\item CV_XGB - logical variable, use cross-validation for gradient boosting? Default: TRUE.
\item nfolds - number of folds in cross-validation, default: nfolds = 5.
\item nrounds - the max number of iterations, default: nrounds = 50.
}}
}
\value{
List with following parameters:
\item{y_hat_out_of_sample}{imputed out of sample outcomes}
\item{y_hat_sample}{predicted sample data using fitted model}
\item{outcome_fit}{fitted model to impute the data}
}
\description{
Internal generic function to impute values of y
}
\section{Methods (by class)}{
\itemize{
\item \code{OR}: Function to estimate imputation_y using EBLUP

\item \code{IPW}: Function to estimate imputation_y using EBLUP

\item \code{EBLUP}: Function to estimate imputation_y using EBLUP

\item \code{MQ}: Function to estimate imputation_y using MQ

\item \code{RF}: Function to estimate imputation_y using RF

\item \code{XGB}: Function to estimate imputation_y using RF
}}

