% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_y.R
\name{mutate_obj_fit}
\alias{mutate_obj_fit}
\alias{mutate_obj_fit.EBLUP}
\alias{mutate_obj_fit.MQ}
\alias{mutate_obj_fit.RF}
\alias{mutate_obj_fit.XGB}
\title{Generic function to format data}
\usage{
mutate_obj_fit(...)

\method{mutate_obj_fit}{EBLUP}(obj_fit_y, model_formula, ...)

\method{mutate_obj_fit}{MQ}(obj_fit_y, model_formula, ...)

\method{mutate_obj_fit}{RF}(obj_fit_y, model_formula, ...)

\method{mutate_obj_fit}{XGB}(obj_fit_y, model_formula, ...)
}
\arguments{
\item{...}{Additional parameters.}

\item{obj_fit_y}{Object to fit outcome y}

\item{model_formula}{Model formula.}
}
\description{
Generic function to format data
}
\section{Methods (by class)}{
\itemize{
\item \code{mutate_obj_fit(EBLUP)}: Function to estimate fit_y using EBLUP

\item \code{mutate_obj_fit(MQ)}: Function to estimate fit_y using MQ

\item \code{mutate_obj_fit(RF)}: Function to estimate fit_y using RF

\item \code{mutate_obj_fit(XGB)}: Function to estimate fit_y using XGB

}}
