% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{obtain_IPW_data}
\alias{obtain_IPW_data}
\title{Obtain IPW data}
\usage{
obtain_IPW_data(obj_hte, params_p_score, params_impute_y, estimated_p_score)
}
\arguments{
\item{obj_hte}{Object to estimate hte}

\item{params_p_score}{List with parameters to fit propensity score:
\itemize{
\item model_formula
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
}}

\item{params_impute_y}{List with parameters to fit imputation model:
\itemize{
\item model_formula
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
\item type_model - type of outcome.
}}

\item{estimated_p_score}{Vector of estimated propensity scores, default: \code{estimated_p_score = NULL}.}
}
\value{
Estimates of propensity score.
}
\description{
Internal function to obtain data to fit IPW and NIPW estimators
}
