% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{outcome_regression}
\alias{outcome_regression}
\title{Outcome regression}
\usage{
outcome_regression(
  formula_y,
  data_sample,
  data_out_of_sample,
  method_y,
  family = "gaussian",
  method = "continuous",
  tune_RF
)
}
\arguments{
\item{data_sample}{Data from the sample.}

\item{data_out_of_sample}{Data from the remaining part of population.}

\item{family}{Specification for the model link function in glmer()}

\item{method}{Specification for the model link function in mquantreg()}

\item{tune_RF}{Tune parameters in random forest? Default = FALSE.}
}
\description{
Fit outcome regression model
}
