% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{p_score}
\alias{p_score}
\title{Propensity score}
\usage{
p_score(
  formula_p_score,
  data_sample,
  data_out_of_sample = NULL,
  method_p_score
)
}
\arguments{
\item{formula_p_score}{Formula for fitting outcome regression}

\item{data_sample}{Data from the sample}

\item{data_out_of_sample}{Data from the population}

\item{method_p_score}{Method to predict/fit propnsity score}
}
\description{
Fit propensity score model
}
