% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_bootstrap.R
\name{residual_bootstrap}
\alias{residual_bootstrap}
\title{Residual bootstrap}
\usage{
residual_bootstrap(
  y,
  y_hat,
  data_sample,
  n_boot = 250,
  type_boot = c("br1", "br2", "br3", "br4", "br5"),
  boot_seed = 10,
  rand_clust = FALSE,
  ...
)
}
\arguments{
\item{y}{Vector of true outcomes}

\item{y_hat}{Vector of estimated outcomes}

\item{data_sample}{Sample data with group indicator}

\item{n_boot}{Number of bootstrap samples}

\item{type_boot}{Type of bootstrap: \code{br1} - one-level block residual bootstrap,
\code{br2} - two-level block residual bootstrap.}

\item{boot_seed}{Seed to repeat bootstrap sampling}

\item{rand_clust}{Random clusters}

\item{...}{Additional parameters}
}
\description{
One-level and two-level random block residual bootstrap for variance calculations
}
