% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_var.R
\name{sample_bootstrap_indices}
\alias{sample_bootstrap_indices}
\title{Sample bootstrap indices}
\usage{
sample_bootstrap_indices(
  sample_sizes,
  out_of_sample_sizes,
  type_boot,
  n_boot = 50,
  seed = 1
)
}
\arguments{
\item{sample_sizes}{Data from the sample}

\item{out_of_sample_sizes}{Data from the population}

\item{type_boot}{Type of bootstrap sampling}

\item{n_boot}{Number of bootstrap samples}

\item{seed}{Seed to run simulations.}
}
\value{
List with following parameters:
\item{ind_population}{Vector with population indices}
\item{ind_sample}{Vector with sample indices}
}
\description{
Sample bootstrap indices
}
\examples{

m = 50
ni = rep(5, m)
Ni = rep(100, m)
N = sum(Ni)
n = sum(ni)

X <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "unif",
 seed = 1
)


X_outcome <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "lognorm",
 seed = 1
)


populations <- generate_pop(X, X_outcome,
coeffs = get_default_coeffs(),
errors_outcome = get_default_errors_outcome(),
rand_eff_outcome = get_default_rand_eff_outcome(),
rand_eff_p_score = get_default_rand_eff_p_score(),
regression_type = "continuous",
Ni_size  = 100,
m = 50,
no_sim = 1,
seed = 1)

samples <- generate_sample(populations, ni_size = 10,
                           sample_part = "sampled",
                           get_index = TRUE)

data_sample <- data.frame(samples[[1]]$samp_data)
index_sample <- samples[[1]]$index_s
data_out_of_sample <- populations[-index_sample, ]

sample_sizes = as.data.frame(table(data_sample$group))$Freq
out_of_sample_sizes = as.data.frame(table(data_out_of_sample$group))$Freq

bootstrap_indices <- sample_bootstrap_indices(sample_sizes,
                                                out_of_sample_sizes,
                                                type_boot = "sample",
                                                n_boot = 500,
                                                seed = 1)
}
