% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_bootstrap.R
\name{sample_bresiduals}
\alias{sample_bresiduals}
\title{Sample bootstrap residuals}
\usage{
sample_bresiduals(sample_size, n_boot, swr = TRUE, seed = 10)
}
\arguments{
\item{sample_size}{Sample sizes}

\item{n_boot}{Number of bootstrap samples}

\item{swr}{Simple random sampling with replacement? Default = \code{swr = TRUE}.}

\item{seed}{Seed to run simulations.}
}
\value{
List with indices.
}
\description{
Sample bootstrap residuals
}
