% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_subpopulations.R
\name{sample_subpopulations}
\alias{sample_subpopulations}
\title{Sample subpopulations}
\usage{
sample_subpopulations(one_pop, frac_nc, frac_nt, seed = 1)
}
\arguments{
\item{one_pop}{Population to sample from}

\item{frac_nc}{Fraction of controls in each subpopulation}

\item{frac_nt}{Fraction of treated in each subpopulation}

\item{seed}{Seed to repeat the simulations}
}
\description{
Sample subpopulations
}
\examples{

m = 50
ni = rep(5, m)
Ni = rep(100, m)
N = sum(Ni)
n = sum(ni)

X <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "unif",
 seed = 1
)

X_outcome <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "lognorm",
 seed = 1
)

populations <- generate_pop(X, X_outcome,
coeffs = get_default_coeffs(),
errors_outcome = get_default_errors_outcome(),
rand_eff_outcome = get_default_rand_eff_outcome(),
rand_eff_p_score = get_default_rand_eff_p_score(),
regression_type = "continuous",
Ni_size  = 100,
m = 50,
no_sim = 1,
seed = 1)

subpopulation <- sample_subpopulations(populations,
frac_nc = 0.1, frac_nt = 0.1, seed = 1)

}
