% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tau.R
\name{calculate_tau}
\alias{calculate_tau}
\title{Calculate true causal effect}
\usage{
calculate_tau(populations, type_tau = c("regular", "adaptive"))
}
\arguments{
\item{populations}{List with one or several populations}

\item{type_tau}{Type of computation, \code{regular}: constant weights,
\code{adaptive}: adaptive weights}
}
\value{
\item{tau_true}{List with taus accross small areas}
}
\description{
This function calculates causal area effect tau
}
