% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{estimate_hte}
\alias{estimate_hte}
\title{Estimate hte}
\usage{
estimate_hte(...)
}
\arguments{
\item{...}{Additional parameters}

\item{obj_hte}{Object to estimate hte}
}
\value{
List with following parameters:
\item{y_hat_out_of_sample}{imputed out of sample outcomes}
\item{y_hat_sample}{predicted sample data using fitted model}
\item{outcome_fit}{fitted model to impute the data}
}
\description{
Internal generic function to estimate hte
}
