% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_OR.R
\name{fit_OR}
\alias{fit_OR}
\title{Fit outcome regression models}
\usage{
fit_OR(obj_hte, params_OR, params_bootstrap)
}
\arguments{
\item{obj_hte}{Object to estimate hte}

\item{params_OR}{List with parameters to fit outcome regression model:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
\item type_model - type of outcome.
}}

\item{params_bootstrap}{List with parameters to obtain bootstrap variance:
\itemize{
\item boot_var = FALSE,
\item n_boot = 500,
\item boot_seed = 10,
}}
}
\value{
Data frame with outcome regression models.
}
\description{
Function to obtain outcome regession models
}
