% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{gen_outcome}
\alias{gen_outcome}
\alias{gen_outcome.continuous}
\alias{gen_outcome.binary}
\alias{gen_outcome.poisson}
\title{Generate one population}
\usage{
gen_outcome(...)

\method{gen_outcome}{continuous}(
  regression_obj,
  Xreg_outcome,
  errors_outcome,
  coeffs,
  A,
  Ni,
  m,
  rand_eff_outcome,
  seed = 1,
  ...
)

\method{gen_outcome}{binary}(
  regression_obj,
  Xreg_outcome,
  errors_outcome,
  coeffs,
  A,
  Ni,
  m,
  rand_eff_outcome,
  seed = 1,
  ...
)

\method{gen_outcome}{poisson}(
  regression_obj,
  Xreg_outcome,
  errors_outcome,
  coeffs,
  A,
  Ni,
  m,
  rand_eff_outcome,
  seed = 1,
  ...
)
}
\arguments{
\item{...}{Additional parameters}

\item{regression_obj}{Object defining regression type}

\item{Xreg_outcome}{Outcome covariates times coefficients \code{X beta} (without intercept)}

\item{errors_outcome}{List of parameters to generate errors, that is:
\itemize{
\item var_e - variance of errors,
\item mean_e - mean of errors,
\item frac_out - fraction of outlying errors,
\item var_e_out - variance of outlying errors,
\item mean_e_out - mean of outlying errors,
\item disturbance_outcome - shift parameter to generate outlying observations.
}}

\item{coeffs}{List of coefficients for generating outcomes and propensity scores, that is:
\itemize{
\item intercept_outcome - intercept for outcome regression model.
\item intercept_p_score - intercept for propensity score model.
\item coef_outcome - coefficients for outcome regression model.
\item coef_p_score - coefficients for propensity score model.
\item mean_A - mean of coefficients of treatments across different areas.
\item var_A - variance of coefficients of treatments across different areas.
}}

\item{A}{Treatment status}

\item{Ni}{Number of units in areas}

\item{m}{Number of subpopulations.}

\item{rand_eff_outcome}{List of parameters to generate random effects for outcome model, that is:
\itemize{
\item var_re - variance of random effects,
\item mean_re - mean of random effects,
\item frac_out - fraction of outlying random effects,
\item var_re_out - variance of outlying random effects,
\item mean_re_out - mean of outlying random effects.
}}

\item{seed}{Seed to replicate simulations.}
}
\value{
A data frame with generated population
}
\description{
This is a generic function to generate outcome variables
}
\section{Methods (by class)}{
\itemize{
\item \code{continuous}: Generate population with normally distributed outcomes

\item \code{binary}: Generate population with Poisson distributed outcomes

\item \code{poisson}: Generate population with binomial distributed outcomes
}}

