% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{generate_pop}
\alias{generate_pop}
\title{Generate populations}
\usage{
generate_pop(
  X = matrix(),
  X_outcome = NULL,
  X_p_score = NULL,
  coeffs = get_default_coeffs(),
  errors_outcome = get_default_errors_outcome(),
  rand_eff_outcome = get_default_rand_eff_outcome(),
  rand_eff_p_score = get_default_rand_eff_p_score(),
  regression_type = c("continuous", "binary", "poisson", "nb"),
  fct_coef = c("sin", "exp", "sin_exp", "lin"),
  Ni_size = 100,
  m = 50,
  no_sim = 100,
  seed = 1
)
}
\arguments{
\item{X}{Covariates for generating outcome and propensity score.}

\item{X_outcome}{Covariates for generating outcome.}

\item{X_p_score}{Covariates for generating propensity score.}

\item{coeffs}{List of coefficients for generating outcomes and propensity scores, that is:
\itemize{
\item intercept_outcome - intercept for outcome regression model.
\item intercept_p_score - intercept for propensity score model.
\item coef_outcome - coefficients for outcome regression model.
\item coef_p_score - coefficients for propensity score model.
\item mean_A - mean of coefficients of treatments across different areas.
\item var_A - variance of coefficients of treatments across different areas.
}}

\item{errors_outcome}{List of parameters to generate errors, that is:
\itemize{
\item var_e - variance of errors,
\item mean_e - mean of errors,
\item frac_out - fraction of outlying errors,
\item var_e_out - variance of outlying errors,
\item mean_e_out - mean of outlying errors,
\item disturbance_outcome - shift parameter to generate outlying observations.
}}

\item{rand_eff_outcome}{List of parameters to generate random effects for outcome model, that is:
\itemize{
\item var_re - variance of random effects,
\item mean_re - mean of random effects,
\item frac_out - fraction of outlying random effects,
\item var_re_out - variance of outlying random effects,
\item mean_re_out - mean of outlying random effects.
}}

\item{rand_eff_p_score}{List of parameters to generate random effects for propensity score model, that is:
\itemize{
\item var_re - variance of random effects,
\item mean_re - mean of random effects,
\item frac_out - fraction of outlying random effects,
\item var_re_out - variance of outlying random effects,
\item mean_re_out - mean of outlying random effects.
}}

\item{regression_type}{Type of outcomes.}

\item{fct_coef}{Functional form of covariates.}

\item{Ni_size}{Vector of subpopulations sizes.}

\item{m}{Number of subpopulations.}

\item{no_sim}{Number of simulations.}

\item{seed}{Seed to replicate simulations.}
}
\value{
List composed of data frames which contain covariates X,
treatment status A, group indicator group, propensity score p_score, and outcome y
}
\description{
Function to generate populations using parametric model assumptions.
}
\examples{

m = 50
ni = rep(5, m)
Ni = rep(100, m)
N = sum(Ni)
n = sum(ni)

X <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "unif",
 seed = 1
)


X_outcome <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "lognorm",
 seed = 1
)


populations <- generate_pop(X, X_outcome,
coeffs = get_default_coeffs(),
errors_outcome = get_default_errors_outcome(),
rand_eff_outcome = get_default_rand_eff_outcome(),
rand_eff_p_score = get_default_rand_eff_p_score(),
regression_type = "continuous",
Ni_size  = 100,
m = 50,
fct_coef = "sin_exp",
no_sim = 1,
seed = 1)

}
