% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample.R
\name{generate_sample}
\alias{generate_sample}
\title{Sample from the population}
\usage{
generate_sample(
  populations,
  ni_size = 5,
  sample_part = "sampled",
  get_index = F
)
}
\arguments{
\item{populations}{List with populations.}

\item{ni_size}{Sample size.}

\item{sample_part}{Select sampled, non_sampled or both parts.}

\item{get_index}{Get the index of sampled observations
or population, default: \code{sample_part = sampled}.}
}
\description{
Sample from the population
}
\examples{

m = 50
ni = rep(5, m)
Ni = rep(100, m)
N = sum(Ni)
n = sum(ni)

X <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "unif",
 seed = 1
)


X_outcome <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "lognorm",
 seed = 1
)


populations <- generate_pop(X, X_outcome,
coeffs = get_default_coeffs(),
errors_outcome = get_default_errors_outcome(),
rand_eff_outcome = get_default_rand_eff_outcome(),
rand_eff_p_score = get_default_rand_eff_p_score(),
regression_type = "continuous",
Ni_size  = 100,
m = 50,
no_sim = 10,
seed = 1)

samples <- generate_sample(populations,  ni_size = 5,
sample_part = "sampled")

}
