% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{hte}
\alias{hte}
\title{Heterogenous treatment effect}
\usage{
hte(
  formula_y,
  formula_p_score,
  type_tau,
  data_sample,
  data_out_of_sample,
  method_y = c("EBLUP", "MQ", "RF"),
  method_p_score = c("EBLUP", "MQ", "RF"),
  p_score_estimate = NULL,
  tune_RF = F,
  ...
)
}
\arguments{
\item{data_sample}{Data from the sample.}

\item{data_out_of_sample}{Data from the remaining part of population.}

\item{p_score_estimate}{Vector of estimated propensity scores. Default: \code{p_score_estimate = NULL}.}

\item{tune_RF}{Tune parameters in random forest? Default = FALSE.}

\item{...}{Additional parameters}

\item{type_hte}{Type of estimation method, choose between: \code{OR}, \code{IPW},
\code{NIPW}, \code{AIPW}.}

\item{params_p_score}{List with parameters to fit propensity score:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
}}

\item{params_impute_y}{List with parameters to fit imputation model:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
\item type_model - type of outcome.
}}

\item{params_OR}{List with parameters to fit outcome regression model:
\itemize{
\item model_formula - model formula,
\item method - estimation method, choose between: \code{EBLUP},
\code{MQ}, \code{RF}, \code{XGB},
\item tune_RF - tune random forest parameters,
\item xgboost_params - list of parameters to obtain predictions
using gradient boosting, default: \code{CV_XGB = TRUE},
\code{nfolds = 5}, \code{nrounds = 50}.
\item type_model - type of outcome.
}}
}
\description{
Calculates heterogrenous treatment effects for subpopulations
}
\examples{

#' @examples

m = 50
ni = rep(10, m)
Ni = rep(200, m)
N = sum(Ni)
n = sum(ni)

X <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "unif",
 seed = 1
)

X_outcome <- generate_X(
 n = N,
 p = 1,
 covariance_norm = NULL,
 cov_type = "lognorm",
 seed = 1
)

populations <- generate_pop(X, X_outcome,
coeffs = get_default_coeffs(),
errors_outcome = get_default_errors_outcome(),
rand_eff_outcome = get_default_rand_eff_outcome(),
rand_eff_p_score = get_default_rand_eff_p_score(),
regression_type = "continuous",
Ni_size  = 200,
m = 50,
no_sim = 1,
seed = 10)

samples <- generate_sample(populations, ni_size = 10,
                           sample_part = "sampled",
                           get_index = TRUE)

data_sample <- data.frame(samples[[1]]$samp_data)
index_sample <- samples[[1]]$index_s
data_out_of_sample <- populations[-index_sample, ]

model_formula_OR = y ~ X1 + Xo1 + (1|group)


}
