% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte.R
\name{outcome_regression}
\alias{outcome_regression}
\title{Outcome regression}
\usage{
outcome_regression(
  formula_y,
  data_sample,
  data_out_of_sample,
  method_y,
  family = "gaussian",
  method = "continuous",
  tune_RF
)
}
\arguments{
\item{formula_y}{Formula for fitting outcome regression}

\item{data_sample}{Data from the sample}

\item{data_out_of_sample}{Data from the population}

\item{method_y}{Method to predict/fit the outcome}

\item{family}{Specification for the model link function in glmer()}

\item{method}{Specification for the model link function in mquantreg()}

\item{tune_RF}{Tune random forest parameters}
}
\description{
Fit outcome regression model
}
