% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mq.R
\name{predict.mquantfit}
\alias{predict.mquantfit}
\title{Making predictions for mquantfit}
\usage{
\method{predict}{mquantfit}(
  object,
  newdata = object$data,
  regression_type = c("continuous", "binary", "nb", "poisson"),
  ...
)
}
\arguments{
\item{object}{A \code{gab} object}

\item{newdata}{New data (default is the training data for \code{object})}

\item{regression_type}{Type of regression "continuous", "binary", "nb", "poisson"}

\item{...}{Additional parameters.}
}
\value{
a matrix, each row corresponds to one iteration.
}
\description{
Making predictions for mquantfit
}
