% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_bootstrap.R
\name{resid_boot}
\alias{resid_boot}
\alias{resid_boot.br1}
\alias{resid_boot.br2}
\title{Bootstrap residuals}
\usage{
resid_boot(...)

\method{resid_boot}{br1}(obj_boot, ...)

\method{resid_boot}{br2}(obj_boot, ...)
}
\arguments{
\item{...}{Additional parameters}

\item{obj_boot}{Object to estimate residual bootstrap}
}
\description{
Internal generic function to bootstrap residuals
}
\section{Methods (by class)}{
\itemize{
\item \code{br1}: One-level block residual bootstrap

\item \code{br2}: Two-level block residual bootstrap
}}

