% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_subpopulation.R
\name{sample_subpopulations}
\alias{sample_subpopulations}
\title{Sample subpopulations}
\usage{
sample_subpopulations(one_pop, frac_nc, frac_nt)
}
\arguments{
\item{one_pop}{Population to sample from}

\item{frac_nc}{Fraction of controls in each subpopulation}

\item{frac_nt}{Fraction of treated in each subpopulation}
}
\description{
Sample subpopulations
}
